requirements.txt
falcon
cherrypy
happybase
kafka-python